MVP!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
It works!!
There's a lot of random things to add from here on.
For example, it should reorder the tables according to the foreign
keys (since I got some errors about that while testing).
Also, I need to think of what to do about the default tables laravel
creates for you. For now, I just put them in the schema and deleted
their default migrations, but now the schema looks waaay too big and
hard to look through because of all that noise.